feat (backend): create db_engine module with database drivers, ORM config, engines, and models
- Added database drivers and ORM configuration
- Defined database engines
- Included database models
