+++ agenta-backend/agenta_backend/models/db_engine/models.py
@@ -1,0 +1,69 @@
+from agenta_backend.utils.common import isCloudEE
+
+if isCloudEE():
+    from agenta_backend.commons.models.db_models import (
+        APIKeyDB,
+        WorkspaceDB,
+        OrganizationDB,
+        InvitationDB,
+        UserOrganizationDB,
+        WorkspaceMemberDB,
+        AppDB_ as AppDB,
+        UserDB_ as UserDB,
+        ImageDB_ as ImageDB,
+        TestSetDB_ as TestSetDB,
+        AppVariantDB_ as AppVariantDB,
+        EvaluationDB_ as EvaluationDB,
+        DeploymentDB_ as DeploymentDB,
+        VariantBaseDB_ as VariantBaseDB,
+        AppEnvironmentDB_ as AppEnvironmentDB,
+        AppEnvironmentRevisionDB_ as AppEnvironmentRevisionDB,
+        EvaluatorConfigDB_ as EvaluatorConfigDB,
+        HumanEvaluationDB_ as HumanEvaluationDB,
+        EvaluationScenarioDB_ as EvaluationScenarioDB,
+        HumanEvaluationScenarioDB_ as HumanEvaluationScenarioDB,
+    )
+else:
+    from agenta_backend.models.db_models import (
+        AppDB,
+        UserDB,
+        ImageDB,
+        TestSetDB,
+        EvaluationDB,
+        DeploymentDB,
+        AppVariantDB,
+        VariantBaseDB,
+        AppEnvironmentDB,
+        AppEnvironmentRevisionDB,
+        EvaluatorConfigDB,
+        HumanEvaluationDB,
+        EvaluationScenarioDB,
+        HumanEvaluationScenarioDB,
+    )
+
+from agenta_backend.models.db_models import (
+    TemplateDB,
+    AppVariantRevisionsDB,
+)
+
+models = [
+    AppDB,
+    UserDB,
+    ImageDB,
+    TestSetDB,
+    TemplateDB,
+    AppVariantDB,
+    DeploymentDB,
+    EvaluationDB,
+    VariantBaseDB,
+    AppEnvironmentDB,
+    AppEnvironmentRevisionDB,
+    EvaluatorConfigDB,
+    HumanEvaluationDB,
+    EvaluationScenarioDB,
+    AppVariantRevisionsDB,
+    HumanEvaluationScenarioDB,
+]
+
+if isCloudEE():
+    models.extend([OrganizationDB, WorkspaceDB, APIKeyDB, InvitationDB, UserOrganizationDB, WorkspaceMemberDB])  # type: ignore
